requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
motor==3.6.0
priority==2.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.2
python-dotenv==1.0.1
pytz==2024.2
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
wsproto==1.2.0
